requirements.txt
fastapi
uvicorn
google-generativeai
pinecone
python-dotenv
pydantic
